+++ coreactivity.php
@@ -46,11 +46,11 @@ define( 'COREACTIVITY_URL', $coreactivity_urlname_basic );
 define( 'COREACTIVITY_D4PLIB_PATH', $coreactivity_dirname_basic . 'd4plib/' );
 define( 'COREACTIVITY_D4PLIB_URL', $coreactivity_urlname_basic . 'd4plib/' );
 
-require_once( COREACTIVITY_D4PLIB_PATH . 'core.php' );
+require_once COREACTIVITY_D4PLIB_PATH . 'core.php';
 
-require_once( COREACTIVITY_PATH . 'core/autoload.php' );
-require_once( COREACTIVITY_PATH . 'core/bridge.php' );
-require_once( COREACTIVITY_PATH . 'core/functions.php' );
+require_once COREACTIVITY_PATH . 'core/autoload.php';
+require_once COREACTIVITY_PATH . 'core/bridge.php';
+require_once COREACTIVITY_PATH . 'core/functions.php';
 
 coreactivity();
 coreactivity_settings();