+++ forms/content-dashboard.php
@@ -9,10 +9,10 @@ if ( ! defined( 'ABSPATH' ) ) {
     <div class="d4p-cards-wrapper">
 		<?php
 
-		include( COREACTIVITY_PATH . 'forms/content-dashboard-overall.php' );
-		include( COREACTIVITY_PATH . 'forms/content-dashboard-database.php' );
-		include( COREACTIVITY_PATH . 'forms/content-dashboard-statistics.php' );
-		include( COREACTIVITY_PATH . 'forms/content-dashboard-location.php' );
+		include COREACTIVITY_PATH . 'forms/content-dashboard-overall.php';
+		include COREACTIVITY_PATH . 'forms/content-dashboard-database.php';
+		include COREACTIVITY_PATH . 'forms/content-dashboard-statistics.php';
+		include COREACTIVITY_PATH . 'forms/content-dashboard-location.php';
 
 		?>
     </div>