+++ forms/content-update.php
@@ -13,8 +13,8 @@ if ( ! defined( 'ABSPATH' ) ) {
         <div class="d4p-update-info">
 			<?php
 
-			include( COREACTIVITY_PATH . 'forms/setup-database.php' );
-			include( COREACTIVITY_PATH . 'forms/setup-statistics.php' );
+			include COREACTIVITY_PATH . 'forms/setup-database.php';
+			include COREACTIVITY_PATH . 'forms/setup-statistics.php';
 
 			coreactivity()->schedule_geo_db_update();